Secure endpoints by checking requestor email

+++ src/graphql/resolver/post/mutation.ts
@@ -1,8 +1,15 @@
 import { PostCreateRequest } from '../../../dynamodb/model/post-model';
-import postService from '../../../service/post-service';
+import postService, { getPostById } from '../../../service/post-service';
 
-export const createPost = async (parent, { request }: { request: PostCreateRequest }) => {
+export const createPost = async (parent, { request }: { request: PostCreateRequest }, context) => {
   console.log(`Creating post for ${request.userId}`);
+  if (context.userId !== request.userId) {
+    return {
+      message: `The user ${context.userId as string} is not authenticated to access this resource`,
+      success: false,
+    };
+  }
+
   try {
     const post = await postService.createPost(request);
     return {
@@ -18,8 +25,17 @@ export const createPost = async (parent, { request }: { request: PostCreateReque
   }
 };
 
-export const deletePost = async (parent, { id }: { id: string }) => {
+export const deletePost = async (parent, { id }: { id: string }, context) => {
   console.log(`Deleting post ${id}`);
+
+  const post = await getPostById(id);
+  if (context.userId !== post.userId) {
+    return {
+      message: `The user ${context.userId as string} is not authenticated to access this resource`,
+      success: false,
+    };
+  }
+
   try {
     await postService.deletePost(id);
     return {
